Change the jface snippet for Notification Popup
Refactoring
- use widget factories for all widgets
- use builder api for notification popup
- rename snippet

New functionality
- use fadeIn for notification which is more natural

Issue: https://github.com/eclipse-platform/eclipse.platform.ui/issues/1226
